Refactor + add tree counting average value.
