+++ README.md
@@ -63,4 +63,3 @@ TODO list (with no priority):
 - Custom allocators checks.
 - Add doxygen documentation.
 - Add more classes with operations which are often used with segment tree.
-- Test a combine operation option which supports taking not lengths of segments but their borders.
