+++ include/curried/dynamic_max_segment_tree.hpp
