+++ include/curried/dynamic_min_segment_tree.hpp
