+++ include/curried/dynamic_negate_segment_tree.hpp
