+++ include/curried/dynamic_sum_segment_tree.hpp
@@ -19,7 +19,7 @@ using DynamicSumSegmentTree =
         ValueT,
         SumT,
         decltype([](const SumT& val1, const SumT& val2) -> SumT { return val1 + val2; }),
-        decltype([](const ValueT& val, KeyT length) -> SumT { return val * length; }),
+        decltype([](const ValueT& val, KeyT begin, KeyT end) -> SumT { return val * (end - begin); }),
         UpdateOp,
         UpdateArgT,
         Allocator
