+++ include/impl/dynamic_segment_tree_range_get_variation_base.hpp
@@ -40,25 +40,6 @@ protected:
 template <class KeyT,
           class ValueT,
           class GetValueT,
-          dst::conc::ValueAndLengthGetInitializer<
-              ValueT, KeyT, GetValueT> SegGetInit>
-class DynamicSegmentTreeRangeGetInitVariationBase<KeyT, ValueT,
-                                                  GetValueT, SegGetInit> {
-protected:
-    DynamicSegmentTreeRangeGetInitVariationBase(const SegGetInit& segInit)
-        : _segInitializer(segInit) {}
-protected:
-
-    GetValueT _initGet(KeyT begin, KeyT end, const ValueT& val) const {
-        return this->_segInitializer(val, end - begin);
-    }
-protected:
-    const SegGetInit _segInitializer;
-};
-
-template <class KeyT,
-          class ValueT,
-          class GetValueT,
           dst::conc::ValueAndBordersGetInitializer<
               ValueT, KeyT, GetValueT> SegGetInit>
 class DynamicSegmentTreeRangeGetInitVariationBase<KeyT, ValueT,
@@ -101,28 +82,6 @@ protected:
                           [[maybe_unused]] KeyT /*separation*/,
                           [[maybe_unused]] KeyT /*rightEnd*/) const {
         return _segCombiner(leftValue, rightValue);
-    }
-protected:
-    const SegGetComb _segCombiner;
-};
-
-template <class KeyT,
-          class GetValueT,
-          dst::conc::ValueAndLengthGetCombiner<GetValueT, KeyT> SegGetComb>
-class DynamicSegmentTreeRangeGetCombineVariationBase<KeyT, GetValueT,
-                                                     SegGetComb> {
-protected:
-    DynamicSegmentTreeRangeGetCombineVariationBase(const SegGetComb& segComb)
-        : _segCombiner(segComb) {}
-protected:
-    GetValueT _combineGet(const GetValueT& leftValue,
-                          const GetValueT& rightValue,
-                          KeyT leftBegin,
-                          KeyT separation,
-                          KeyT rightEnd) const {
-        const KeyT leftLength = separation - leftBegin;
-        const KeyT rightLength = rightEnd - separation;
-        return _segCombiner(leftValue, rightValue, leftLength, rightLength);
     }
 protected:
     const SegGetComb _segCombiner;
