+++ include/impl/node_base.hpp
