+++ test/test_max.cpp
@@ -1,36 +1,44 @@
 #include <gtest/gtest.h>
-#include <dynamic_min_segment_tree.hpp>
+#include <curried/dynamic_max_segment_tree.hpp>
 
 #include <random>
-#include "reference/min_seg_tree_reference.hpp"
+#include "reference/max_seg_tree_reference.hpp"
 
-using dst::DynamicMinSegmentTree;
+using dst::DynamicMaxSegmentTree;
 
-TEST(DynamicMinSegmentTree, Construct) {
-    [[maybe_unused]] auto tree = DynamicMinSegmentTree<int, int>(0, 42, 34);
+TEST(DynamicMaxSegmentTree, Construct) {
+    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 34);
 }
 
-TEST(DynamicMinSegmentTree, SimpleRangeGet) {
-    auto tree = DynamicMinSegmentTree<int, int>(0, 42, 34);
+TEST(DynamicMaxSegmentTree, SimpleRangeGet) {
+    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 34);
     EXPECT_EQ(tree.rangeGet(5, 17), 34);
 }
 
-TEST(DynamicMinSegmentTree, RangeGetAfterUpdate) {
-    auto tree = DynamicMinSegmentTree<int, int, int, std::plus<int>>(0, 42, 34);
+TEST(DynamicMaxSegmentTree, RangeGetAfterUpdate) {
+    auto tree = DynamicMaxSegmentTree<int, int, int, std::plus<int>>(0, 42, 34);
     tree.update(12, 22, 4);
-    EXPECT_EQ(tree.rangeGet(5, 17), 34);
+    EXPECT_EQ(tree.rangeGet(5, 17), 34 + 4);
     EXPECT_EQ(tree.rangeGet(12, 18), 34 + 4);
 }
 
-TEST(DynamicMinSegmentTree, RangeGetAfterSet) {
-    auto tree = DynamicMinSegmentTree<int, int>(0, 42, 34);
+TEST(DynamicMaxSegmentTree, RangeGetAfterSet) {
+    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 34);
     tree.set(12, 22, 4);
     EXPECT_EQ(tree.rangeGet(13, 20), 4);
-    EXPECT_EQ(tree.rangeGet(2, 35), 4);
+    EXPECT_EQ(tree.rangeGet(2, 35), 34);
 }
 
-TEST(DynamicMinSegmentTree, LadderUpRight) {
-    auto tree = DynamicMinSegmentTree<int, int>(0, 42, 0);
+TEST(DynamicMaxSegmentTree, UpdateAndSet) {
+    auto tree = DynamicMaxSegmentTree<int, int, int, std::plus<int>>(0, 42, 34);
+    tree.update(12, 22, 4);
+    tree.set(17, 27, 66);
+    EXPECT_EQ(tree.rangeGet(5, 17), 34 + 4);
+    EXPECT_EQ(tree.rangeGet(12, 18), 66);
+}
+
+TEST(DynamicMaxSegmentTree, LadderUpRight) {
+    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 0);
     tree.set(41, 42, 10000000);
     tree.set(40, 41, 1000000);
     tree.set(39, 40, 100000);
@@ -40,15 +48,15 @@ TEST(DynamicMinSegmentTree, LadderUpRight) {
     tree.set(35, 36, 10);
     tree.set(34, 35, 1);
 
-    EXPECT_EQ(tree.rangeGet(0, 42), 0);
-    EXPECT_EQ(tree.rangeGet(36, 42), 100);
-    EXPECT_EQ(tree.rangeGet(0, 36), 0);
-    EXPECT_EQ(tree.rangeGet(39, 42), 100000);
-    EXPECT_EQ(tree.rangeGet(0, 39), 0);
+    EXPECT_EQ(tree.rangeGet(0, 42), 10000000);
+    EXPECT_EQ(tree.rangeGet(36, 42), 10000000);
+    EXPECT_EQ(tree.rangeGet(0, 36), 10);
+    EXPECT_EQ(tree.rangeGet(39, 42), 10000000);
+    EXPECT_EQ(tree.rangeGet(0, 39), 10000);
 }
 
-TEST(DynamicMinSegmentTree, LadderUpLeft) {
-    auto tree = DynamicMinSegmentTree<int, int>(0, 42, 0);
+TEST(DynamicMaxSegmentTree, LadderUpLeft) {
+    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 0);
     tree.set(0, 1, 10000000);
     tree.set(1, 2, 1000000);
     tree.set(2, 3, 100000);
@@ -58,15 +66,15 @@ TEST(DynamicMinSegmentTree, LadderUpLeft) {
     tree.set(6, 7, 10);
     tree.set(7, 7, 1);
 
-    EXPECT_EQ(tree.rangeGet(0, 42), 0);
-    EXPECT_EQ(tree.rangeGet(0, 6), 100);
-    EXPECT_EQ(tree.rangeGet(6, 42), 0);
-    EXPECT_EQ(tree.rangeGet(0, 3), 100000);
-    EXPECT_EQ(tree.rangeGet(3, 42), 0);
+    EXPECT_EQ(tree.rangeGet(0, 42), 10000000);
+    EXPECT_EQ(tree.rangeGet(0, 6), 10000000);
+    EXPECT_EQ(tree.rangeGet(6, 42), 10);
+    EXPECT_EQ(tree.rangeGet(0, 3), 10000000);
+    EXPECT_EQ(tree.rangeGet(3, 42), 10000);
 }
 
-TEST(DynamicMinSegmentTree, LadderDownRight) {
-    auto tree = DynamicMinSegmentTree<int, int>(0, 42, 0);
+TEST(DynamicMaxSegmentTree, LadderDownRight) {
+    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 0);
     tree.set(41, 42, -10000000);
     tree.set(40, 41, -1000000);
     tree.set(39, 40, -100000);
@@ -76,15 +84,15 @@ TEST(DynamicMinSegmentTree, LadderDownRight) {
     tree.set(35, 36, -10);
     tree.set(34, 35, -1);
 
-    EXPECT_EQ(tree.rangeGet(0, 42), -10000000);
-    EXPECT_EQ(tree.rangeGet(36, 42), -10000000);
-    EXPECT_EQ(tree.rangeGet(0, 36), -10);
-    EXPECT_EQ(tree.rangeGet(39, 42), -10000000);
-    EXPECT_EQ(tree.rangeGet(0, 39),  -10000);
+    EXPECT_EQ(tree.rangeGet(0, 42), 0);
+    EXPECT_EQ(tree.rangeGet(36, 42), -100);
+    EXPECT_EQ(tree.rangeGet(0, 36), 0);
+    EXPECT_EQ(tree.rangeGet(39, 42), -100000);
+    EXPECT_EQ(tree.rangeGet(0, 39),  0);
 }
 
-TEST(DynamicMinSegmentTree, LadderDownLeft) {
-    auto tree = DynamicMinSegmentTree<int, int>(0, 42, 0);
+TEST(DynamicMaxSegmentTree, LadderDownLeft) {
+    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 0);
     tree.set(0, 1, -10000000);
     tree.set(1, 2, -1000000);
     tree.set(2, 3, -100000);
@@ -94,38 +102,16 @@ TEST(DynamicMinSegmentTree, LadderDownLeft) {
     tree.set(6, 7, -10);
     tree.set(7, 7, -1);
 
-    EXPECT_EQ(tree.rangeGet(0, 42), -10000000);
-    EXPECT_EQ(tree.rangeGet(0, 6), -10000000);
-    EXPECT_EQ(tree.rangeGet(6, 42), -10);
-    EXPECT_EQ(tree.rangeGet(0, 3), -10000000);
-    EXPECT_EQ(tree.rangeGet(3, 42), -10000);
+    EXPECT_EQ(tree.rangeGet(0, 42), 0);
+    EXPECT_EQ(tree.rangeGet(0, 6), -100);
+    EXPECT_EQ(tree.rangeGet(6, 42), 0);
+    EXPECT_EQ(tree.rangeGet(0, 3), -100000);
+    EXPECT_EQ(tree.rangeGet(3, 42), 0);
 }
 
-TEST(DynamicMinSegmentTree, FuzzTestSetGet) {
-    auto tree = DynamicMinSegmentTree<int, int>(0, 1000, 0);
-    auto reference = MinSegTreeReference<int, int>(0, 1000, 0);
-
-    std::mt19937 generator(42);
-
-    for (std::size_t i = 0; i < 100; ++i) {
-        std::size_t rngStart = generator() % 500; // [0..500)
-        std::size_t rngLen = generator() % 500;   // [0..500)
-        int setVal = generator() % 1000; // [0..100)
-        tree.set(rngStart, rngStart + rngLen, setVal);
-        reference.set(rngStart, rngStart + rngLen, setVal);
-    }
-
-    for (std::size_t i = 0; i < 50; ++i) {
-        int idx = generator() % 100; // [0..1000)
-        auto treeRes = tree.get(idx);
-        auto refRes = reference.get(idx);
-        EXPECT_EQ(treeRes, refRes);
-    }
-}
-
-TEST(DynamicMinSegmentTree, FuzzTestSetUpdateGet) {
-    auto tree = DynamicMinSegmentTree<int, int, int, std::plus<int>>(0, 1000, 0);
-    auto reference = MinSegTreeReference<int, int>(0, 1000, 0);
+TEST(DynamicMaxSegmentTree, FuzzTestSetUpdateGet) {
+    auto tree = DynamicMaxSegmentTree<int, int, int, std::plus<int>>(0, 1000, 0);
+    auto reference = MaxSegTreeReference<int, int>(0, 1000, 0);
 
     std::mt19937 generator(42);
 
@@ -155,9 +141,9 @@ TEST(DynamicMinSegmentTree, FuzzTestSetUpdateGet) {
     }
 }
 
-TEST(DynamicMinSegmentTree, FuzzTestMixedSetUpdateGet) {
-    auto tree = DynamicMinSegmentTree<int, int, int, std::plus<int>>(0, 1000, 0);
-    auto reference = MinSegTreeReference<int, int>(0, 1000, 0);
+TEST(DynamicMaxSegmentTree, FuzzTestMixedSetUpdateGet) {
+    auto tree = DynamicMaxSegmentTree<int, int, int, std::plus<int>>(0, 1000, 0);
+    auto reference = MaxSegTreeReference<int, int>(0, 1000, 0);
 
     std::mt19937 generator(54);
 
@@ -186,10 +172,9 @@ TEST(DynamicMinSegmentTree, FuzzTestMixedSetUpdateGet) {
     }
 }
 
-
-TEST(DynamicMinSegmentTree, FuzzTestSetRangeGet) {
-    auto tree = DynamicMinSegmentTree<int, int>(0, 1000, 0);
-    auto reference = MinSegTreeReference<int, int>(0, 1000, 0);
+TEST(DynamicMaxSegmentTree, FuzzTestSetRangeGet) {
+    auto tree = DynamicMaxSegmentTree<int, int>(0, 1000, 0);
+    auto reference = MaxSegTreeReference<int, int>(0, 1000, 0);
 
     std::mt19937 generator(42);
