+++ test/test_min.cpp
@@ -1,44 +1,36 @@
 #include <gtest/gtest.h>
-#include <dynamic_max_segment_tree.hpp>
+#include <curried/dynamic_min_segment_tree.hpp>
 
 #include <random>
-#include "reference/max_seg_tree_reference.hpp"
+#include "reference/min_seg_tree_reference.hpp"
 
-using dst::DynamicMaxSegmentTree;
+using dst::DynamicMinSegmentTree;
 
-TEST(DynamicMaxSegmentTree, Construct) {
-    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 34);
+TEST(DynamicMinSegmentTree, Construct) {
+    [[maybe_unused]] auto tree = DynamicMinSegmentTree<int, int>(0, 42, 34);
 }
 
-TEST(DynamicMaxSegmentTree, SimpleRangeGet) {
-    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 34);
+TEST(DynamicMinSegmentTree, SimpleRangeGet) {
+    auto tree = DynamicMinSegmentTree<int, int>(0, 42, 34);
     EXPECT_EQ(tree.rangeGet(5, 17), 34);
 }
 
-TEST(DynamicMaxSegmentTree, RangeGetAfterUpdate) {
-    auto tree = DynamicMaxSegmentTree<int, int, int, std::plus<int>>(0, 42, 34);
+TEST(DynamicMinSegmentTree, RangeGetAfterUpdate) {
+    auto tree = DynamicMinSegmentTree<int, int, int, std::plus<int>>(0, 42, 34);
     tree.update(12, 22, 4);
-    EXPECT_EQ(tree.rangeGet(5, 17), 34 + 4);
+    EXPECT_EQ(tree.rangeGet(5, 17), 34);
     EXPECT_EQ(tree.rangeGet(12, 18), 34 + 4);
 }
 
-TEST(DynamicMaxSegmentTree, RangeGetAfterSet) {
-    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 34);
+TEST(DynamicMinSegmentTree, RangeGetAfterSet) {
+    auto tree = DynamicMinSegmentTree<int, int>(0, 42, 34);
     tree.set(12, 22, 4);
     EXPECT_EQ(tree.rangeGet(13, 20), 4);
-    EXPECT_EQ(tree.rangeGet(2, 35), 34);
+    EXPECT_EQ(tree.rangeGet(2, 35), 4);
 }
 
-TEST(DynamicMaxSegmentTree, UpdateAndSet) {
-    auto tree = DynamicMaxSegmentTree<int, int, int, std::plus<int>>(0, 42, 34);
-    tree.update(12, 22, 4);
-    tree.set(17, 27, 66);
-    EXPECT_EQ(tree.rangeGet(5, 17), 34 + 4);
-    EXPECT_EQ(tree.rangeGet(12, 18), 66);
-}
-
-TEST(DynamicMaxSegmentTree, LadderUpRight) {
-    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 0);
+TEST(DynamicMinSegmentTree, LadderUpRight) {
+    auto tree = DynamicMinSegmentTree<int, int>(0, 42, 0);
     tree.set(41, 42, 10000000);
     tree.set(40, 41, 1000000);
     tree.set(39, 40, 100000);
@@ -48,15 +40,15 @@ TEST(DynamicMaxSegmentTree, LadderUpRight) {
     tree.set(35, 36, 10);
     tree.set(34, 35, 1);
 
-    EXPECT_EQ(tree.rangeGet(0, 42), 10000000);
-    EXPECT_EQ(tree.rangeGet(36, 42), 10000000);
-    EXPECT_EQ(tree.rangeGet(0, 36), 10);
-    EXPECT_EQ(tree.rangeGet(39, 42), 10000000);
-    EXPECT_EQ(tree.rangeGet(0, 39), 10000);
+    EXPECT_EQ(tree.rangeGet(0, 42), 0);
+    EXPECT_EQ(tree.rangeGet(36, 42), 100);
+    EXPECT_EQ(tree.rangeGet(0, 36), 0);
+    EXPECT_EQ(tree.rangeGet(39, 42), 100000);
+    EXPECT_EQ(tree.rangeGet(0, 39), 0);
 }
 
-TEST(DynamicMaxSegmentTree, LadderUpLeft) {
-    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 0);
+TEST(DynamicMinSegmentTree, LadderUpLeft) {
+    auto tree = DynamicMinSegmentTree<int, int>(0, 42, 0);
     tree.set(0, 1, 10000000);
     tree.set(1, 2, 1000000);
     tree.set(2, 3, 100000);
@@ -66,15 +58,15 @@ TEST(DynamicMaxSegmentTree, LadderUpLeft) {
     tree.set(6, 7, 10);
     tree.set(7, 7, 1);
 
-    EXPECT_EQ(tree.rangeGet(0, 42), 10000000);
-    EXPECT_EQ(tree.rangeGet(0, 6), 10000000);
-    EXPECT_EQ(tree.rangeGet(6, 42), 10);
-    EXPECT_EQ(tree.rangeGet(0, 3), 10000000);
-    EXPECT_EQ(tree.rangeGet(3, 42), 10000);
+    EXPECT_EQ(tree.rangeGet(0, 42), 0);
+    EXPECT_EQ(tree.rangeGet(0, 6), 100);
+    EXPECT_EQ(tree.rangeGet(6, 42), 0);
+    EXPECT_EQ(tree.rangeGet(0, 3), 100000);
+    EXPECT_EQ(tree.rangeGet(3, 42), 0);
 }
 
-TEST(DynamicMaxSegmentTree, LadderDownRight) {
-    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 0);
+TEST(DynamicMinSegmentTree, LadderDownRight) {
+    auto tree = DynamicMinSegmentTree<int, int>(0, 42, 0);
     tree.set(41, 42, -10000000);
     tree.set(40, 41, -1000000);
     tree.set(39, 40, -100000);
@@ -84,15 +76,15 @@ TEST(DynamicMaxSegmentTree, LadderDownRight) {
     tree.set(35, 36, -10);
     tree.set(34, 35, -1);
 
-    EXPECT_EQ(tree.rangeGet(0, 42), 0);
-    EXPECT_EQ(tree.rangeGet(36, 42), -100);
-    EXPECT_EQ(tree.rangeGet(0, 36), 0);
-    EXPECT_EQ(tree.rangeGet(39, 42), -100000);
-    EXPECT_EQ(tree.rangeGet(0, 39),  0);
+    EXPECT_EQ(tree.rangeGet(0, 42), -10000000);
+    EXPECT_EQ(tree.rangeGet(36, 42), -10000000);
+    EXPECT_EQ(tree.rangeGet(0, 36), -10);
+    EXPECT_EQ(tree.rangeGet(39, 42), -10000000);
+    EXPECT_EQ(tree.rangeGet(0, 39),  -10000);
 }
 
-TEST(DynamicMaxSegmentTree, LadderDownLeft) {
-    auto tree = DynamicMaxSegmentTree<int, int>(0, 42, 0);
+TEST(DynamicMinSegmentTree, LadderDownLeft) {
+    auto tree = DynamicMinSegmentTree<int, int>(0, 42, 0);
     tree.set(0, 1, -10000000);
     tree.set(1, 2, -1000000);
     tree.set(2, 3, -100000);
@@ -102,38 +94,16 @@ TEST(DynamicMaxSegmentTree, LadderDownLeft) {
     tree.set(6, 7, -10);
     tree.set(7, 7, -1);
 
-    EXPECT_EQ(tree.rangeGet(0, 42), 0);
-    EXPECT_EQ(tree.rangeGet(0, 6), -100);
-    EXPECT_EQ(tree.rangeGet(6, 42), 0);
-    EXPECT_EQ(tree.rangeGet(0, 3), -100000);
-    EXPECT_EQ(tree.rangeGet(3, 42), 0);
+    EXPECT_EQ(tree.rangeGet(0, 42), -10000000);
+    EXPECT_EQ(tree.rangeGet(0, 6), -10000000);
+    EXPECT_EQ(tree.rangeGet(6, 42), -10);
+    EXPECT_EQ(tree.rangeGet(0, 3), -10000000);
+    EXPECT_EQ(tree.rangeGet(3, 42), -10000);
 }
 
-TEST(DynamicMaxSegmentTree, FuzzTestSetGet) {
-    auto tree = DynamicMaxSegmentTree<int, int>(0, 1000, 0);
-    auto reference = MaxSegTreeReference<int, int>(0, 1000, 0);
-
-    std::mt19937 generator(42);
-
-    for (std::size_t i = 0; i < 100; ++i) {
-        std::size_t rngStart = generator() % 500; // [0..500)
-        std::size_t rngLen = generator() % 500;   // [0..500)
-        int setVal = generator() % 1000; // [0..100)
-        tree.set(rngStart, rngStart + rngLen, setVal);
-        reference.set(rngStart, rngStart + rngLen, setVal);
-    }
-
-    for (std::size_t i = 0; i < 50; ++i) {
-        int idx = generator() % 100; // [0..1000)
-        auto treeRes = tree.get(idx);
-        auto refRes = reference.get(idx);
-        EXPECT_EQ(treeRes, refRes);
-    }
-}
-
-TEST(DynamicMaxSegmentTree, FuzzTestSetUpdateGet) {
-    auto tree = DynamicMaxSegmentTree<int, int, int, std::plus<int>>(0, 1000, 0);
-    auto reference = MaxSegTreeReference<int, int>(0, 1000, 0);
+TEST(DynamicMinSegmentTree, FuzzTestSetUpdateGet) {
+    auto tree = DynamicMinSegmentTree<int, int, int, std::plus<int>>(0, 1000, 0);
+    auto reference = MinSegTreeReference<int, int>(0, 1000, 0);
 
     std::mt19937 generator(42);
 
@@ -163,9 +133,9 @@ TEST(DynamicMaxSegmentTree, FuzzTestSetUpdateGet) {
     }
 }
 
-TEST(DynamicMaxSegmentTree, FuzzTestMixedSetUpdateGet) {
-    auto tree = DynamicMaxSegmentTree<int, int, int, std::plus<int>>(0, 1000, 0);
-    auto reference = MaxSegTreeReference<int, int>(0, 1000, 0);
+TEST(DynamicMinSegmentTree, FuzzTestMixedSetUpdateGet) {
+    auto tree = DynamicMinSegmentTree<int, int, int, std::plus<int>>(0, 1000, 0);
+    auto reference = MinSegTreeReference<int, int>(0, 1000, 0);
 
     std::mt19937 generator(54);
 
@@ -194,16 +164,17 @@ TEST(DynamicMaxSegmentTree, FuzzTestMixedSetUpdateGet) {
     }
 }
 
-TEST(DynamicMaxSegmentTree, FuzzTestSetRangeGet) {
-    auto tree = DynamicMaxSegmentTree<int, int>(0, 1000, 0);
-    auto reference = MaxSegTreeReference<int, int>(0, 1000, 0);
+
+TEST(DynamicMinSegmentTree, FuzzTestSetRangeGet) {
+    auto tree = DynamicMinSegmentTree<int, int>(0, 1000, 0);
+    auto reference = MinSegTreeReference<int, int>(0, 1000, 0);
 
     std::mt19937 generator(42);
 
     for (std::size_t i = 0; i < 100; ++i) {
         std::size_t rngStart = generator() % 500; // [0..500)
         std::size_t rngLen = generator() % 500;   // [0..500)
-        int setVal = generator() % 1000; // [0..100)
+        int setVal = generator() % 1000; // [0..1000)
         tree.set(rngStart, rngStart + rngLen, setVal);
         reference.set(rngStart, rngStart + rngLen, setVal);
     }
