+++ test/test_negate.cpp
@@ -2,7 +2,7 @@
 
 #include <random>
 
-#include <dynamic_negate_segment_tree.hpp>
+#include <curried/dynamic_negate_segment_tree.hpp>
 #include "reference/seg_tree_reference_base.hpp"
 
 template <std::integral KeyT, class ValueT, class Allocator = std::allocator<ValueT>>
