+++ test/test_simple.cpp
@@ -1,6 +1,8 @@
 #include <gtest/gtest.h>
+#include <random>
+#include "reference/seg_tree_reference_base.hpp"
 
-#include <simple_get_set_dynamic_segment_tree.hpp>
+#include <curried/simple_get_set_dynamic_segment_tree.hpp>
 
 using dst::SimpleGetSetDynamicSegmentTree;
 
@@ -19,4 +21,49 @@ TEST(SimpleGetSetDynamicSegmentTree, RangeSet) {
     EXPECT_EQ(tree.get(17), 77);
     EXPECT_EQ(tree.get(8), 77);
     EXPECT_EQ(tree.get(37), 77);
+}
+
+TEST(SimpleGetSetDynamicSegmentTree, FuzzTestSetGet) {
+    auto tree = SimpleGetSetDynamicSegmentTree<int, int>(0, 1000, 0);
+    auto reference = SegTreeReferenceBase<int, int>(0, 1000, 0);
+
+    std::mt19937 generator(42);
+
+    for (std::size_t i = 0; i < 100; ++i) {
+        std::size_t rngStart = generator() % 500; // [0..500)
+        std::size_t rngLen = generator() % 500;   // [0..500)
+        int setVal = generator() % 1000; // [0..100)
+        tree.set(rngStart, rngStart + rngLen, setVal);
+        reference.set(rngStart, rngStart + rngLen, setVal);
+    }
+
+    for (std::size_t i = 0; i < 50; ++i) {
+        int idx = generator() % 1000; // [0..1000)
+        auto treeRes = tree.get(idx);
+        auto refRes = reference.get(idx);
+        EXPECT_EQ(treeRes, refRes);
+    }
+}
+
+TEST(SimpleGetSetDynamicSegmentTree, FuzzTestSetGetMixed) {
+    auto tree = SimpleGetSetDynamicSegmentTree<int, int>(0, 1000, 0);
+    auto reference = SegTreeReferenceBase<int, int>(0, 1000, 0);
+
+    std::mt19937 generator(42);
+
+    for (std::size_t i = 0; i < 100; ++i) {
+        std::size_t rngStart = generator() % 500; // [0..500)
+        std::size_t rngLen = generator() % 500;   // [0..500)
+
+        if (int opType = generator() % 2; opType == 0) {
+            int setVal = generator() % 1000; // [0..1000)
+            tree.set(rngStart, rngStart + rngLen, setVal);
+            reference.set(rngStart, rngStart + rngLen, setVal);
+        } else {
+            int idx = generator() % 1000; // [0..1000)
+            auto treeRes = tree.get(idx);
+            auto refRes = reference.get(idx);
+            EXPECT_EQ(treeRes, refRes);
+        }
+    }
 }
